% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneSigTest.R
\name{GeneSigTest}
\alias{GeneSigTest}
\title{Perform Gene Expression Significance Test}
\usage{
GeneSigTest(
  Bulk,
  frac,
  ref,
  nMC = 1000,
  p_threshold = 0.05,
  refine = FALSE,
  auto = TRUE
)
}
\arguments{
\item{Bulk}{Bulk expression matrix}

\item{frac}{Estimated cell type fractions}

\item{ref}{Reference profile matrix}

\item{nMC}{the number of bootstrapping times}

\item{p_threshold}{the threshold of pvalue to determine the significant gene}

\item{refine}{whether perform refinement}

\item{auto}{automatically determine whether perform refinement}

\item{covariate}{The data.frame object contains the covariate information of each sample}
}
\value{
A list object contain two object
 call: A binary matrix (row: gene, column: cell type) indicating which gene has high confidence to be accurated estimated
 pval: A p-value matrix (row: gene, column: cell type) indicating the confidence
}
\description{
Perform Gene Expression Significance Test
}
